requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
astunparse==1.6.3
blinker==1.6.2
blis==0.7.9
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.7
Flask==2.3.3
flask-cors==4.0.0
flatbuffers==22.12.6
gast==0.4.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
langcodes==3.3.0
language_data==1.3.0
libclang==15.0.6.1
marisa-trie==1.1.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml_dtypes==0.2.0
murmurhash==1.0.9
namex==0.0.8
numpy==1.23.5
opencv-python==4.6.0.66
opt_einsum==3.3.0
optree==0.10.0
packaging==23.2
pathlib==1.0.1
pillow==9.5.0
preshed==3.0.6
protobuf==3.19.6
pydantic==1.10.13
Pygments==2.16.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorflow==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.3.0
thinc==8.1.10
tqdm==4.66.2
typer==0.9.0
typing-inspection==0.4.0
# REMOVED typing_extensions==4.7.1 → let pip resolve best match
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
Werkzeug==2.3.7
wrapt==1.14.1
